requirements.txt
django==5.0.0
djangorestframework==3.15.2
boto3==1.35.54
mysqlclient==2.2.5
PyMySQL==1.1.1
cryptography==43.0.3
django-environ==0.11.2
pytest==8.3.3
pytest-django==4.9.0
moto==5.0.20